refactor: removed pyi file
